Fix magnet stalling when sweeping
Magnet did not stall when ramping. Cause of problem was not identified
so this should be considered as a workaround.
Symptom of problem: When a field value is reached by the sensitivity
condition the magnet axis never returns to HOLD. Thus, a new ramp can be
initiated but never starts even though the magnet state is RTOS (ramp to
set).